cpufreq: ZenX: don't use cpu_is_offline 'unlikely' branch prediction in the timer is probably going to cost us several cycles since it will be wrong very often especially when the device is idle.
cpu_is_offline is defined as unlikely(!cpu_online). Since we expect
auxillary CPUs to be off when the device is idle (likely fairly often),
replace it with !cpu_online and not do any branch prediction.

Not sure the performance improvement this is, but we definitely
don't need the unlikely macro here.